added custom hook useStartLevel
